ya funcionan y devuelven todo loans y returns

+++ src/main/java/com/example/demo/repository/LoanRepository.java
@@ -1,12 +1,9 @@
 package com.example.demo.repository;
 
 import java.io.Serializable;
-import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import com.example.demo.entity.Book;
@@ -16,7 +13,7 @@ import com.example.demo.entity.User;
 @Repository("loanRepository")
 public interface LoanRepository extends JpaRepository<Loan, Serializable> {
 
-	Book findById(int id);
+	Book findById(long id);
 
 	int countByUserAndDeletedFalse(User user);
